Removed feeling lucky setting and Refactored code Feeling lucky was kinda dumb, so I removed it from 2.0 Removed some warnings from Utils Moved Utils to the utils package
