+++ app/src/main/res/xml/preferences.xml
@@ -1,20 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
     <PreferenceCategory android:title="Preferences">
-        <SwitchPreference
-            android:defaultValue="false"
-            android:key="pref_luckyKorean"
-            android:summary="@string/lucky_kor_false"
-            android:title="I'm Feeling Lucky - Korean" />
-        <SwitchPreference
-            android:defaultValue="false"
-            android:key="pref_luckyEnglish"
-            android:summary="@string/lucky_eng_false"
-            android:title="I'm Feeling Lucky - English" />
         <Preference
             android:defaultValue="3"
             android:key="pref_fav_count"
-            android:summary="You have 3 favorites"
+            tools:summary="You have 3 favorites"
             android:title="Favorites">
             <intent android:targetClass="com.a494studios.koreanconjugator.settings.FavoritesActivity"
                 android:targetPackage="com.a494studios.koreanconjugator"/>
